fix: move to mjs files instead of js. Do some clean up.
