+++ index.html
@@ -5,7 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <link rel="stylesheet" href="style.css">
     <title>Virtual keyboard</title>
-    <script src="script.js" type="module"></script>
+    <script src="script.mjs" type="module"></script>
 </head>
 <body>
 </body>
